+++ src/main/groovy/cloneable/App.groovy
@@ -40,6 +40,9 @@ class App implements Callable<Integer> {
     @Option(names = ["-b", "--skip-local-bare-repos"], description = "If a bare repository directory exists locally (the name of the repo ending with '.git'), then it will not be printed out.  This is useful for cloning only missing repositories.")
     Boolean skipLocalBare = false
 
+    @Option(names = ["-m", "--match-topics"], description = "Require all repositories to have one of the listed topics.  -m can be specified multiple times.")
+    List<String> matchAnyTopic = []
+
     static void main(String... args) {
         int exitCode = new CommandLine(new App()).execute(args);
         System.exit(exitCode);